Formatted files which was responsible for authentication and managing authentication, user registration.
Cars Controller: Managed car inventory (CRUD for Cars)
Created service and implementation for using in carController.

POST: /cars - add a new car
GET: /cars - get a list of cars
GET: /cars/ - get car's detailed information
PUT/PATCH: /cars/ - update car (also manage inventory)
DELETE: /cars/ - delete car
