--- src/main/java/mate/academy/car_service/CarServiceApplication.java
@@ -1,13 +1,0 @@
-package mate.academy.car_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CarServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CarServiceApplication.class, args);
-	}
-
-}
