--- src/main/java/mate/academy/car_service/auth/dto/UserLoginResponseDto.java
@@ -1,14 +1,0 @@
-package mate.academy.car_service.auth.dto;
-
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Data
-public class UserLoginResponseDto {
-    private String token;
-
-    public UserLoginResponseDto(String token) {
-        this.token = token;
-    }
-}
