--- src/main/java/mate/academy/car_service/auth/dto/UserRegisterResponseDto.java
@@ -1,11 +1,0 @@
-package mate.academy.car_service.auth.dto;
-
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-@Data
-@Accessors(chain = true)
-public class UserRegisterResponseDto {
-    private Long id;
-    private String email;
-}
