--- src/main/java/mate/academy/car_service/config/MapperConfig.java
@@ -1,14 +1,0 @@
-package mate.academy.car_service.config;
-
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.NullValueCheckStrategy;
-
-@org.mapstruct.MapperConfig(
-        componentModel = "spring",
-        injectionStrategy = InjectionStrategy.CONSTRUCTOR,
-        nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS,
-        implementationPackage = "<PACKAGE_NAME>.impl"
-
-)
-public class MapperConfig {
-}
