--- src/main/java/mate/academy/car_service/exception/RegistrationException.java
@@ -1,7 +1,0 @@
-package mate.academy.car_service.exception;
-
-public class RegistrationException extends Exception {
-    public RegistrationException(String message) {
-        super(message);
-    }
-}
