--- src/main/java/mate/academy/car_service/model/Role.java
@@ -1,31 +1,0 @@
-package mate.academy.car_service.model;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.Accessors;
-import org.springframework.security.core.GrantedAuthority;
-
-@Entity
-@Table(name = "roles")
-@Getter
-@Setter
-@Accessors(chain = true)
-public class Role implements GrantedAuthority {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(name = "role_name", columnDefinition = "varchar", unique = true, nullable = false)
-    @Enumerated(value = EnumType.STRING)
-    private RoleName roleName;
-
-    @Override
-    public String getAuthority() {
-        return roleName.name();
-    }
-
-    public enum RoleName {
-        ROLE_MANAGER,
-        ROLE_CUSTOMER
-    }
-}
