--- src/main/java/mate/academy/car_service/repo/RoleRepository.java
@@ -1,8 +1,0 @@
-package mate.academy.car_service.repo;
-
-import mate.academy.car_service.model.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Role findByRoleName(Role.RoleName roleName);
-}
