--- src/main/java/mate/academy/car_service/security/AuthenticationService.java
@@ -1,25 +1,0 @@
-package mate.academy.car_service.security;
-
-import lombok.RequiredArgsConstructor;
-import mate.academy.car_service.auth.dto.UserLoginRequestDto;
-import mate.academy.car_service.auth.dto.UserLoginResponseDto;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AuthenticationService {
-    private final JwtUtil jwtUtil;
-    private final AuthenticationManager authenticationManager;
-
-    public UserLoginResponseDto authenticate(UserLoginRequestDto requestDto) {
-        final Authentication authentication = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        requestDto.getEmail(), requestDto.getPassword())
-        );
-        String token = jwtUtil.generateToken(authentication.getName());
-        return new UserLoginResponseDto(token);
-    }
-}
