--- src/main/java/mate/academy/car_service/security/CustomUserDetailsService.java
@@ -1,20 +1,0 @@
-package mate.academy.car_service.security;
-
-import lombok.RequiredArgsConstructor;
-import mate.academy.car_service.repo.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return userRepository.findByEmail(email).orElseThrow(() ->
-                new UsernameNotFoundException("Can't find carsharingservice.user by email: " + email));
-    }
-}
