--- src/main/java/mate/academy/car_service/security/JwtAuthenticationFilter.java
@@ -1,49 +1,0 @@
-package mate.academy.car_service.security;
-
-import jakarta.servlet.*;
-import jakarta.servlet.http.*;
-import java.io.IOException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-@Component
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-  private final JwtUtil jwtUtil;
-  private final UserDetailsService userDetailsService;
-
-  @Override
-  protected void doFilterInternal(
-    HttpServletRequest request,
-    HttpServletResponse response,
-    FilterChain filterChain
-  ) throws ServletException, IOException {
-    String token = getToken(request);
-
-    if (token != null && jwtUtil.isValidToken(token)) {
-      String username = jwtUtil.getUsername(token);
-      UserDetails userDetails = userDetailsService.loadUserByUsername(username);
-      Authentication authentication = new UsernamePasswordAuthenticationToken(
-        userDetails, null, userDetails.getAuthorities()
-      );
-      SecurityContextHolder.getContext().setAuthentication(authentication);
-    }
-
-    filterChain.doFilter(request, response);
-  }
-
-  private String getToken(HttpServletRequest request) {
-    String bearerToken = request.getHeader("Authorization");
-    if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-      return bearerToken.substring(7);
-    }
-    return null;
-  }
-}
