+++ src/main/java/mate/academy/carservice/CarServiceApplication.java
@@ -1,0 +1,13 @@
+package mate.academy.carservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CarServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CarServiceApplication.class, args);
+    }
+
+}
