+++ src/main/java/mate/academy/carservice/auth/dto/UpdateRoleRequestDto.java
@@ -1,0 +1,14 @@
+package mate.academy.carservice.auth.dto;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import lombok.Data;
+import lombok.experimental.Accessors;
+
+@Data
+@Accessors(chain = true)
+public class UpdateRoleRequestDto {
+    @NotNull
+    @Positive
+    private Long roleId;
+}
