+++ src/main/java/mate/academy/carservice/auth/dto/UserLoginRequestDto.java
@@ -1,0 +1,12 @@
+package mate.academy.carservice.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Data;
+
+@Data
+public class UserLoginRequestDto {
+    @NotBlank
+    private String email;
+    @NotBlank
+    private String password;
+}
