+++ src/main/java/mate/academy/carservice/auth/dto/UserLoginResponseDto.java
@@ -1,0 +1,14 @@
+package mate.academy.carservice.auth.dto;
+
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Data
+public class UserLoginResponseDto {
+    private String token;
+
+    public UserLoginResponseDto(String token) {
+        this.token = token;
+    }
+}
