+++ src/main/java/mate/academy/carservice/config/MapperConfig.java
@@ -1,0 +1,14 @@
+package mate.academy.carservice.config;
+
+import org.mapstruct.InjectionStrategy;
+import org.mapstruct.NullValueCheckStrategy;
+
+@org.mapstruct.MapperConfig(
+        componentModel = "spring",
+        injectionStrategy = InjectionStrategy.CONSTRUCTOR,
+        nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS,
+        implementationPackage = "<PACKAGE_NAME>.impl"
+
+)
+public class MapperConfig {
+}
