+++ src/main/java/mate/academy/carservice/exception/RegistrationException.java
@@ -1,0 +1,7 @@
+package mate.academy.carservice.exception;
+
+public class RegistrationException extends Exception {
+    public RegistrationException(String message) {
+        super(message);
+    }
+}
