+++ src/main/java/mate/academy/carservice/model/Role.java
@@ -1,0 +1,38 @@
+package mate.academy.carservice.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.Accessors;
+import org.springframework.security.core.GrantedAuthority;
+
+@Entity
+@Table(name = "roles")
+@Getter
+@Setter
+@Accessors(chain = true)
+public class Role implements GrantedAuthority {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(name = "role_name", columnDefinition = "varchar", unique = true, nullable = false)
+    @Enumerated(value = EnumType.STRING)
+    private RoleName roleName;
+
+    @Override
+    public String getAuthority() {
+        return roleName.name();
+    }
+
+    public enum RoleName {
+        ROLE_MANAGER,
+        ROLE_CUSTOMER
+    }
+}
