+++ src/main/java/mate/academy/carservice/repo/RoleRepository.java
@@ -1,0 +1,8 @@
+package mate.academy.carservice.repo;
+
+import mate.academy.carservice.model.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Role findByRoleName(Role.RoleName roleName);
+}
