+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.application.name=car_service
 spring.datasource.url=jdbc:mysql://localhost:3306/car
 spring.datasource.username=root
 spring.datasource.password=156354OlEh
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.open-in-view=false
