--- src/test/java/mate/academy/car_service/CarServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package mate.academy.car_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CarServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
