+++ src/test/java/mate/academy/carservice/CarServiceApplicationTests.java
@@ -1,0 +1,12 @@
+package mate.academy.carservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CarServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+}
