+++ src/test/resources/application.properties
@@ -1,7 +1,10 @@
+spring.application.name=car_service
+spring.datasource.url=jdbc:mysql://localhost:3306/car
+spring.datasource.username=root
+spring.datasource.password=156354OlEh
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.show-sql=true
+spring.jpa.open-in-view=false
 
-
-spring.datasource.url=jdbc:tc:mysql:8:///service
-spring.datasource.username=test
-spring.datasource.password=test
 jwt.expiration=300000
 jwt.secret=secretkey123456121213123124234OlEh
